@@ -36,19 +36,25 @@
   ],
   "main": "./out/src/extension",
   "scripts": {
-    "vscode:prepublish": "node ./node_modules/vscode/bin/compile",
-    "compile": "node ./node_modules/vscode/bin/compile -watch -p ./",
+    "vscode:prepublish": "tsc -p ./",
+    "compile": "tsc -watch -p ./",
     "postinstall": "node ./node_modules/vscode/bin/install",
     "clean": "rimraf out/src/*"
   },
   "devDependencies": {
-    "typescript": "^1.8.5",
-    "vscode": "^0.11.0",
-    "rimraf": "latest",
-    "tslint": "latest"
+    "@types/css": "0.0.30",
+    "@types/lodash.flow": "^3.5.2",
+    "@types/mocha": "^2.2.40",
+    "@types/node": "^6.0.40",
+    "mocha": "^3.2.0",
+    "rimraf": "^2.6.1",
+    "tslint": "^5.1.0",
+    "typescript": "^2.2.2",
+    "vscode": "^1.1.0"
   },
   "dependencies": {
     "css": "^2.2.1",
+    "lodash.flow": "^3.5.0",
     "xxhashjs": "^0.2.1"
   }
 }
