@@ -1,8 +1,8 @@
 import { Uri } from "vscode";
 import { readFile } from "fs";
 
-export default (urisPromise: Thenable<Uri[]>, encoding: string) => {
-    return Promise.resolve(urisPromise).then(uris => {
+export default (uris: Uri[]|Thenable<Uri[]>, encoding: string): Thenable<string[]> => {
+    return Promise.resolve(uris).then(uris => {
         return Promise.all(uris.map(uri => new Promise<string>((resolve, reject) => {
             readFile(uri.fsPath, encoding, (err, data) => {
                 if (err) {
