@@ -5,9 +5,9 @@ import { Uri } from "vscode";
 suite("Test uriFilesReader", () => {
     test("Can't load http protocol", () => {
         const httpsUri = Uri.parse("https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css");
-        return uriFilesReader(Promise.resolve([httpsUri]), "utf8").then(data => {
+        return uriFilesReader([httpsUri], "utf8").then(data => {
             assert(false, "Expected promise to be rejected.");
-        }).catch(err => {
+        },err => {
             assert(err.code === "ENOENT");
         });
     });
